Optimizer: convert the DeadEndBlockDumper pass to a test
